+++ ElvUI_MerathilisUI/Core/Changelog/7.34.lua
@@ -2,7 +2,9 @@ local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[734] = {
 	RELEASE_DATE = "TBD",
-	FIXES = {},
+	FIXES = {
+		"Bags: Categorized Bags - fixed a taint error (\"tried to call the protected function 'UpgradeItem()'\") when upgrading an item at an upgrade NPC.",
+	},
 	NEW = {},
 	IMPROVEMENTS = {},
 }