+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ### Changes
 
--
+-   [Fix]: Bags: Categorized Bags - fixed a taint error ("tried to call the protected function 'UpgradeItem()'") when upgrading an item at an upgrade NPC.